+++ packages/alfa-css/src/value/image/gradient/radial/index.ts
@@ -1,0 +1,1 @@
+export * from "./radial";
