+++ packages/alfa-css/tsconfig.json
@@ -57,7 +57,12 @@
     "src/value/image/gradient/linear/linear.ts",
     "src/value/image/gradient/linear/position.ts",
     "src/value/image/gradient/linear/side.ts",
+    "src/value/image/gradient/radial/circle.ts",
+    "src/value/image/gradient/radial/ellipse.ts",
+    "src/value/image/gradient/radial/extent.ts",
+    "src/value/image/gradient/radial/index.ts",
     "src/value/image/gradient/radial/radial.ts",
+    "src/value/image/gradient/radial/shape.ts",
     "src/value/image/image.ts",
     "src/value/image/index.ts",
     "src/value/image/url.ts",
